refactor talk component
